Refactor

+++ src/systems/camera_system.rs
@@ -17,8 +17,8 @@ impl<'s> System<'s> for RotateCameraSystem {
     fn run(&mut self, (cameras, mut locals, time): Self::SystemData) {
         for (camera, local) in (&cameras, &mut locals).join() {
             let value = time.absolute_time().as_secs() as f64 / 10000.0;
-            local.prepend_translation_x(value.cos());
-            local.prepend_translation_y(value.sin());
+            local.prepend_translation_x(value.cos() as f32);
+            local.prepend_translation_y(value.sin() as f32);
         }
     }
 }
